+++ include/hivemind-bridge/IHiveMindBridge.h
