+++ include/hivemind-bridge/IThreadSafeQueue.h
