+++ include/hivemind-bridge/MessageHandler.h
@@ -2,7 +2,7 @@
 #define HIVE_MIND_BRIDGE_MESSAGEHANDLER_H
 
 #include "Callback.h"
-#include "HiveMindBridge/MessageUtils.h"
+#include "hivemind-bridge/MessageUtils.h"
 #include "IMessageHandler.h"
 #include <cpp-common/ILogger.h>
