+++ include/hivemind-bridge/ThreadSafeQueue.h
