+++ test/integration/SendRequestIntegrationTest.cpp
@@ -2,7 +2,7 @@
 #include "../utils/TCPClient.h"
 #include "../utils/HiveMindBridgeFixture.h"
 #include "../utils/BytesTestData.h"
-#include "HiveMindBridge/HiveMindBridge.h"
+#include "hivemind-bridge/HiveMindBridge.h"
 #include <gmock/gmock.h>
 #include <pheromones/HiveMindHostDeserializer.h>
 #include <pheromones/HiveMindHostSerializer.h>
